pyproject.toml
[project]
name = "video-memory-maker"
version = "0.1.0"
description = "動画ハイライト自動作成ツール"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "mediapipe>=0.10.31",
    "numpy>=2.4.1",
    "opencv-python>=4.13.0.90",
    "pillow>=12.1.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "pytest>=9.0.2",
]
